# backend/docs/apifox_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Life Canvas OS API",
    "description": "\n# Life Canvas OS 后端 API\n\n八维生命平衡系统后端接口，支持个人成长管理、日记记录、AI 洞察等功能。\n\n## 主要功能\n\n- **用户认证**: PIN 码设置、验证、修改\n- **系统管理**: 八维系统（饮食、运动、智力、输出、恢复、资产、连接、环境）\n- **用户配置**: 用户信息、设置、AI 配置\n- **日记管理**: 日记 CRUD、附件、编辑历史\n- **AI 洞察**: 基于用户数据生成个性化洞察\n- **数据管理**: 备份、恢复、导出、导入\n\n## 认证方式\n\n本 API 使用 PIN 码认证。\n\n## 响应格式\n\n所有接口遵循统一的响应格式：\n\n```json\n{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1707219200000\n}\n```\n\n## 错误码\n\n- `200`: 成功\n- `201`: 创建成功\n- `400`: 错误请求\n- `401`: 未授权\n- `404`: 资源不存在\n- `409`: 冲突\n- `422`: 参数验证失败\n- `429`: 请求过于频繁\n- `500`: 服务器错误\n\n## 分页参数\n\n| 参数 | 类型 | 必填 | 默认值 | 说明 |\n|------|------|------|--------|------|\n| page | Integer | 否 | 1 | 页码，从 1 开始 |\n| page_size | Integer | 否 | 20 | 每页数量，最大 100 |\n| sort_by | String | 否 | created_at | 排序字段 |\n| sort_order | String | 否 | desc | 排序方向：asc/desc |\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/pin/setup": {
      "post": {
        "tags": [
          "authentication",
          "authentication"
        ],
        "summary": "设置PIN码",
        "description": "设置 PIN 码（首次）\n\n验证规则：6 位数字",
        "operationId": "setup_pin_api_pin_setup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PinSetupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object"
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pin/verify": {
      "post": {
        "tags": [
          "authentication",
          "authentication"
        ],
        "summary": "验证PIN码",
        "description": "验证 PIN 码",
        "operationId": "verify_pin_code_api_pin_verify_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PinVerifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/PinVerifyResponse"
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/pin/change": {
      "post": {
        "tags": [
          "authentication",
          "authentication"
        ],
        "summary": "修改PIN码",
        "description": "修改 PIN 码",
        "operationId": "change_pin_api_pin_change_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PinChangeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object"
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/data/export": {
      "post": {
        "tags": [
          "data-management",
          "data-management"
        ],
        "summary": "导出数据",
        "description": "导出数据\n\nArgs:\n    format: 导出格式 (json, zip)",
        "operationId": "export_data_api_data_export_post",
        "parameters": [
          {
            "name": "format",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "导出格式 (json, zip)",
              "default": "json",
              "title": "Format"
            },
            "description": "导出格式 (json, zip)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "string",
                      "format": "binary"
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/import": {
      "post": {
        "tags": [
          "data-management",
          "data-management"
        ],
        "summary": "导入数据",
        "description": "导入数据\n\nArgs:\n    backup_path: 备份文件路径\n    verify: 是否验证备份文件",
        "operationId": "import_data_api_data_import_post",
        "parameters": [
          {
            "name": "backup_path",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Backup Path"
            }
          },
          {
            "name": "verify",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Verify"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "imported_at": {
                          "type": "string"
                        }
                      }
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PinChangeRequest": {
        "properties": {
          "old_pin": {
            "type": "string",
            "pattern": "^\\d{6}$",
            "title": "Old Pin",
            "description": "旧 PIN 码"
          },
          "new_pin": {
            "type": "string",
            "pattern": "^\\d{6}$",
            "title": "New Pin",
            "description": "新 PIN 码"
          }
        },
        "type": "object",
        "required": [
          "old_pin",
          "new_pin"
        ],
        "title": "PinChangeRequest",
        "description": "修改 PIN 码请求"
      },
      "PinSetupRequest": {
        "properties": {
          "pin": {
            "type": "string",
            "pattern": "^\\d{6}$",
            "title": "Pin",
            "description": "6 位数字 PIN 码"
          }
        },
        "type": "object",
        "required": [
          "pin"
        ],
        "title": "PinSetupRequest",
        "description": "设置 PIN 码请求"
      },
      "PinVerifyRequest": {
        "properties": {
          "pin": {
            "type": "string",
            "pattern": "^\\d{6}$",
            "title": "Pin",
            "description": "6 位数字 PIN 码"
          }
        },
        "type": "object",
        "required": [
          "pin"
        ],
        "title": "PinVerifyRequest",
        "description": "验证 PIN 码请求"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "PinVerifyResponse": {
        "type": "object",
        "properties": {
          "verified": {
            "type": "boolean"
          },
          "user_id": {
            "type": "integer"
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "开发环境"
    }
  ],
...
}